services/orchestrator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
pyyaml==6.0.3
pydantic==2.12.5
